Rickenmann's tr_cap formula
I added the transport capacity formula by Rickenmann (2001). Currently all partitioning methods can be applied to this formula. This formula is based on the critical discharge Qc. In the loop I have also added variables for the transport capacity for each size class and the critical discharge for each size class (only if Rickenmann's tr cap was chosen). These are also included in the data_output.