Renaming
I rename the new matrix Qbi_incoming in Qbi_pass. It represents what passes throught a reach n. Not what comes into this reach.
Qbi_came (was has arrived from previous time_step, is named Qbi_incoming again because we are used to it now. But this we can see later to decide if we change .