Adding pytest as optional requirement
requirements-optional.txt: This file lists the optional or non-essential dependencies. These packages are not required for the core functionality of the project but might be useful for additional features, development, testing, or extra utilities. It can also contain dependencies needed for specific environments (like testing or local development) but not necessary for all users.