Same results as Vjosa 10 days in v1
At this commit I get the same results as version 1, in the case of Vjosa 10 days (wilcock). The way to compute the velocity is relying on the reach velocity before cascades arrive. Qbi incoming from previous time step is considered, and the incoming layers are equally considered in the active layer. In the new version, the Qbi_incoming from previous time step is already deposited in Vdep, and layers are taken from the top instead. May make some changes in the tr_cap