TransportCapacityCalculator class working
Created a TransportCapacityCalculator class. Vjosa and Po tests are working. The D50/dmi definition in the different formula in correspondance with the different partitionning methods needs to be defined more clearly... The partitionning should be included in the class once understood. The class could link to SedimentarySystem object to avoid ahving its own attributed once such a class is created.