Some changes for calculating time lag
- Cascade having some sediment class with 0 flux are put an elapsed time that is nan (instead of 0). Then we compute the time lag only on non nan elapsed time values, by taking the mean elapsed time of each cascade and then taking the minimum of these times (first cascade arriving).
We also update in the code when the elapsed time is used to sort cascades for example. Instead we use the nanmean as the global elapsed time of the cascade.
This produced some little changes in the order in which cascade are deposited and in turn in the Vdep and thus mobilised volume.
We update the Unit tests.