Update DCASCADE_class.py
Now evolving bed within the width/hypsometric loop, as well as calculating transport. Big assumption - eroding from the deeper areas, depositing at the water's edge. The deposit is mixed in various layers due to the changes in Q, which may be OK. Always eroding starting from the deepest creates an armour layer (good) but may act a bit strange as it searches for finer material.